add test to cover another scenario when we need to get configuration from app.config (in disk)
