+++ src/NHiLo.Tests.Integration/HiLo/Repository/MySql/MySqlIntegrationTest.cs
@@ -3,21 +3,21 @@ using DotNet.Testcontainers.Configurations;
 using DotNet.Testcontainers.Containers;
 using FluentAssertions;
 using Microsoft.Extensions.Configuration;
-using Oracle.ManagedDataAccess.Client;
+using MySql.Data.MySqlClient;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NHiLo.Tests.Integration.Repository.Oracle
+namespace NHiLo.Tests.Integration.HiLo.Repository.MySql
 {
     [Collection("Database Integration")]
-    public class OracleIntegrationTest
+    public class MySqlIntegrationTest
     {
         private readonly ITestOutputHelper _output;
 
-        public OracleIntegrationTest(ITestOutputHelper output)
+        public MySqlIntegrationTest(ITestOutputHelper output)
         {
             _output = output;
         }
@@ -26,24 +26,26 @@ namespace NHiLo.Tests.Integration.Repository.Oracle
         [Trait("Category", "Integration")]
         public async Task Should_ConnectToABrandNewDatabaseAndGetKey()
         {
-            var testcontainersBuilder = new TestcontainersBuilder<OracleTestcontainer>()
-                .WithDatabase(new OracleTestcontainerConfiguration()
+            var testcontainersBuilder = new TestcontainersBuilder<MySqlTestcontainer>()
+                .WithDatabase(new MySqlTestcontainerConfiguration
                 {
-                    Password = "password"
+                    Database = "myDataBase",
+                    Username = "myUser",
+                    Password = "myPassword",
+
                 });
 
             await using (var testcontainer = testcontainersBuilder.Build())
             {
                 await testcontainer.StartAsync();
-
                 var appSettings = $@"{{
                     ""NHiLo"":{{
                         ""DefaultMaxLo"" : ""100""
                     }},
                     ""ConnectionStrings"":{{
                         ""NHiLo"":{{
-                            ""ConnectionString"":""{ testcontainer.ConnectionString }"",
-                            ""ProviderName"":""System.Data.OracleClient""
+                            ""ConnectionString"":""{testcontainer.ConnectionString}"",
+                            ""ProviderName"":""MySql.Data.MySqlClient""
                         }}
                     }}
                 }}";
@@ -51,21 +53,22 @@ namespace NHiLo.Tests.Integration.Repository.Oracle
                 builder.AddJsonStream(new MemoryStream(Encoding.UTF8.GetBytes(appSettings)));
                 var factory = new HiLoGeneratorFactory(builder.Build());
 
-                var generator = factory.GetKeyGenerator("myOracleEntity");
+                var generator = factory.GetKeyGenerator("myMySqlEntity");
                 long key = generator.GetKey();
                 _output.WriteLine($"Key generated: '{key}'");
                 key.Should().BeGreaterThan(0, "is expected the key to be greater than 0.");
 
-                await using (var connection = new OracleConnection(testcontainer.ConnectionString))
+                await using (var connection = new MySqlConnection(testcontainer.ConnectionString))
                 {
                     connection.Open();
-                    await using (var cmd = connection.CreateCommand())
+                    await using (var cmd = new MySqlCommand())
                     {
-                        cmd.CommandText = "SELECT * FROM NHILO WHERE ENTITY = 'myOracleEntity'";
+                        cmd.Connection = connection;
+                        cmd.CommandText = "SELECT * FROM NHILO WHERE ENTITY = 'myMySqlEntity'";
                         using (var reader = cmd.ExecuteReader())
                         {
                             reader.Read();
-                            long nexttHi = reader.GetInt64(reader.GetOrdinal("NEXT_HI"));
+                            long nexttHi = reader.GetInt64("NEXT_HI");
                             _output.WriteLine($"Next Hi value: '{nexttHi}'");
                             nexttHi.Should().Be(2, "is expected the next Hi value to be equal to 2 (first execution).");
                         }
