+++ src/NHiLo.Tests.Integration/HiLo/Repository/Oracle/OracleIntegrationTest.cs
@@ -3,21 +3,21 @@ using DotNet.Testcontainers.Configurations;
 using DotNet.Testcontainers.Containers;
 using FluentAssertions;
 using Microsoft.Extensions.Configuration;
-using MySql.Data.MySqlClient;
+using Oracle.ManagedDataAccess.Client;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NHiLo.Tests.Integration.Repository.MySql
+namespace NHiLo.Tests.Integration.HiLo.Repository.Oracle
 {
     [Collection("Database Integration")]
-    public class MySqlIntegrationTest
+    public class OracleIntegrationTest
     {
         private readonly ITestOutputHelper _output;
 
-        public MySqlIntegrationTest(ITestOutputHelper output)
+        public OracleIntegrationTest(ITestOutputHelper output)
         {
             _output = output;
         }
@@ -26,26 +26,24 @@ namespace NHiLo.Tests.Integration.Repository.MySql
         [Trait("Category", "Integration")]
         public async Task Should_ConnectToABrandNewDatabaseAndGetKey()
         {
-            var testcontainersBuilder = new TestcontainersBuilder<MySqlTestcontainer>()
-                .WithDatabase(new MySqlTestcontainerConfiguration
+            var testcontainersBuilder = new TestcontainersBuilder<OracleTestcontainer>()
+                .WithDatabase(new OracleTestcontainerConfiguration()
                 {
-                    Database = "myDataBase",
-                    Username = "myUser",
-                    Password = "myPassword",
-                    
+                    Password = "password"
                 });
 
             await using (var testcontainer = testcontainersBuilder.Build())
             {
                 await testcontainer.StartAsync();
+
                 var appSettings = $@"{{
                     ""NHiLo"":{{
                         ""DefaultMaxLo"" : ""100""
                     }},
                     ""ConnectionStrings"":{{
                         ""NHiLo"":{{
-                            ""ConnectionString"":""{ testcontainer.ConnectionString }"",
-                            ""ProviderName"":""MySql.Data.MySqlClient""
+                            ""ConnectionString"":""{testcontainer.ConnectionString}"",
+                            ""ProviderName"":""System.Data.OracleClient""
                         }}
                     }}
                 }}";
@@ -53,22 +51,21 @@ namespace NHiLo.Tests.Integration.Repository.MySql
                 builder.AddJsonStream(new MemoryStream(Encoding.UTF8.GetBytes(appSettings)));
                 var factory = new HiLoGeneratorFactory(builder.Build());
 
-                var generator = factory.GetKeyGenerator("myMySqlEntity");
+                var generator = factory.GetKeyGenerator("myOracleEntity");
                 long key = generator.GetKey();
                 _output.WriteLine($"Key generated: '{key}'");
                 key.Should().BeGreaterThan(0, "is expected the key to be greater than 0.");
 
-                await using (var connection = new MySqlConnection(testcontainer.ConnectionString))
+                await using (var connection = new OracleConnection(testcontainer.ConnectionString))
                 {
                     connection.Open();
-                    await using (var cmd = new MySqlCommand())
+                    await using (var cmd = connection.CreateCommand())
                     {
-                        cmd.Connection = connection;
-                        cmd.CommandText = "SELECT * FROM NHILO WHERE ENTITY = 'myMySqlEntity'";
+                        cmd.CommandText = "SELECT * FROM NHILO WHERE ENTITY = 'myOracleEntity'";
                         using (var reader = cmd.ExecuteReader())
                         {
                             reader.Read();
-                            long nexttHi = reader.GetInt64("NEXT_HI");
+                            long nexttHi = reader.GetInt64(reader.GetOrdinal("NEXT_HI"));
                             _output.WriteLine($"Next Hi value: '{nexttHi}'");
                             nexttHi.Should().Be(2, "is expected the next Hi value to be equal to 2 (first execution).");
                         }
