+++ src/NHiLo.Tests.Integration/testhost.dll.config
@@ -5,7 +5,7 @@
 		<section name="nhilo" type="NHiLo.Common.Legacy.KeyGeneratorConfig, NHiLo"/>
 	</configSections>
 	<connectionStrings>
-		<add name="DB1" connectionString="test" providerName="myRepository"/>
+		<add name="DB1" connectionString="test" providerName="NHiLo.InMemory"/>
 	</connectionStrings>
 	<nhilo>
 		<hiloKeyGenerator connectionStringId="DB1" createHiLoStructureIfNotExists="true" defaultMaxLo="10" storageType="Sequence" objectPrefix="SQXX_">
